Change the pms5003 sensor to run constantly
Or rather, change the code to _read_ from it constantly.

By default the sensor just writes data to the serial port every time
it's ready, so if you're not reading it regularly enough it just buffers
and you get out of date data, so the easiest thing to do is just read it
forever and take whatever the latest value is.