Refactor for sensor warmup time
Various sensors on this board need to be warmed up before they start
providing useful data, so I've added a delay to the startup for that

I also kinda massively refactored it to use task cancellation rather
than an event for handling interrupts as I wasn't aware that was an
option when I wrote the event things.

I think the result is kinda nicer?

This also makes sure the pms5003 sensor is actually up and reporting
before continuing which my previous attempt did not, leading to the
first event not including any particulate info